[DEV-47] refactor: 디렉토리 이동
