+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"C:\\app\\codeit_si\\slid-todo\\src\\components\\list-todo.tsx":"1"},{"size":7007,"mtime":1739777344136,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","suppressedMessages":"6","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1nd4jpy","C:\\app\\codeit_si\\slid-todo\\src\\components\\list-todo.tsx",[],[]]
+[{"F:\\download\\slid-todo-client\\src\\app\\page.tsx":"1","F:\\download\\slid-todo-client\\src\\components\\container-recently-todo.tsx":"2","F:\\download\\slid-todo-client\\src\\components\\list-todo.tsx":"3","F:\\download\\slid-todo-client\\src\\containers\\container-recently-todo\\Goal.tsx":"4","F:\\download\\slid-todo-client\\src\\containers\\container-recently-todo\\GoalsListComponent.tsx":"5","F:\\download\\slid-todo-client\\src\\containers\\container-recently-todo\\TodoTitleAndCheckBox.tsx":"6","F:\\download\\slid-todo-client\\src\\types\\container-recently-todo.ts":"7"},{"size":1119,"mtime":1739844049814,"results":"8","hashOfConfig":"9"},{"size":3750,"mtime":1739843280158,"results":"10","hashOfConfig":"9"},{"size":7007,"mtime":1739785160930,"results":"11","hashOfConfig":"9"},{"size":480,"mtime":1739844309305},{"size":723,"mtime":1739844309408},{"size":1123,"mtime":1739844309490},{"size":171,"mtime":1739844310268,"results":"12","hashOfConfig":"9"},{"filePath":"13","messages":"14","suppressedMessages":"15","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"ixuebv",{"filePath":"16","messages":"17","suppressedMessages":"18","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","suppressedMessages":"21","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","suppressedMessages":"24","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"F:\\download\\slid-todo-client\\src\\app\\page.tsx",[],[],"F:\\download\\slid-todo-client\\src\\components\\container-recently-todo.tsx",[],[],"F:\\download\\slid-todo-client\\src\\components\\list-todo.tsx",[],[],"F:\\download\\slid-todo-client\\src\\types\\container-recently-todo.ts",[],[]]
