+++ src/types/container-recently-todo.ts
@@ -1,0 +1,9 @@
+export interface Todo {
+  id: string;
+  title: string;
+  status: "todo" | "done";
+  hasGoal: string | null;
+  hasLink: boolean;
+  hasFile: boolean;
+  createdAt: number;
+}
